moved alert to Home page so it doesn't show on all pages
