+++ client/Routes.js
@@ -4,7 +4,6 @@ import { withRouter, Route, Switch, Redirect } from "react-router-dom";
 import { Login, Signup } from "./components/AuthForm";
 import Home from "./components/Home";
 import { me } from "./store";
-import Survey from "./components/Survey";
 import Tracker from "./components/Tracker";
 import DataVis from "./components/DataVis";
 
@@ -24,7 +23,7 @@ class Routes extends Component {
           <Switch>
             <Route path="/home" component={Home} />
             <Redirect from="/login" to="/home" />
-            <Route path="/survey" component={Survey} />
+
             <Route path="/data" component={DataVis} />
           </Switch>
         ) : (
